parse Android tombstones + iOS PLCrashReport into Sentry-shape events
Native crashes now post a full Sentry-shape structured event in the new
resolved_stack_trace field on both platforms, so the dashboard renders
them with an identical frame-by-frame view (signal metadata, thread
selector, register dump, per-frame image path + instruction address,
debug_meta.images with build ids for symbolication).

Android (vendored SDK, matches co.rivium.trace:rivium-trace-android-sdk 0.2.0):
- Vendor tombstone.proto from AOSP (Apache 2.0, java_package rewritten
  to co.rivium.trace.sdk.tombstone to avoid classloader collision with
  the framework's built-in TombstoneProtos stub).
- Add protobuf-javalite runtime + protobuf-gradle-plugin.
- TombstoneParser: binary Tombstone protobuf -> Sentry-shape JSON + a
  debuggerd-style text fallback for stack_trace.
- NativeCrashReporter: MAX_TRACE_BYTES bumped 64KB -> 512KB (real
  tombstones are 200-400KB), truncation guard skips parsing when the
  stream was cut mid-message, NUL bytes stripped from description /
  processName so Postgres text columns don't reject them (SQLSTATE
  22021).
- RiviumTrace.setupNativeCrashReporting: drain on a background daemon
  thread (avoids NetworkOnMainThreadException), wrap the whole loop
  in try/catch(Throwable) so parser failures cannot crash the host
  app, only advance the watermark on HTTP 200, break on the first
  failure so a transient network error keeps the backlog for the
  next launch instead of purging it.
- RiviumTraceClient.sendErrorSync: log the exception class name so
  silent null-message failures are identifiable in logcat.
- RiviumTraceError: add resolvedStackTrace field (serialized as
  resolved_stack_trace).

iOS (vendored SDK, matches RiviumTrace 0.2.0 pod / SPM):
- NativeCrashReporter: buildSentryShapeJson(report:) maps
  PLCrashReport into the same schema Android emits. All threads with
  per-frame image + instruction_addr, registers on the crashing thread
  only, mechanism with signal + mach_exception + fault_address.
- normalizeSignalName() helper — PLCrashReport.signalInfo.name already
  returns "SIGSEGV", the previous code was double-prefixing and
  producing "SIGSIGSEGV".
- RiviumTraceError: resolvedStackTrace field parallel to Android.

Example app:
- Android: JNI helper in example/android/app/src/main/cpp/ (kept out
  of the plugin so consumers don't inherit an NDK requirement). New
  MethodChannel co.rivium.trace.example/crash_helper on MainActivity
  handles crashNative(kind) with real SIGSEGV / abort / ANR triggers.
- iOS: AppDelegate registers the same channel. Uses
  UnsafeMutableRawPointer(bitPattern: 0x1) for SIGSEGV — Swift traps
  at address 0 via a runtime check that would fatalError before
  PLCrashReporter sees the signal.
- example/lib/main.dart: three new buttons (SIGSEGV / abort / ANR)
  invoke the example-owned channel directly, not via the plugin.

THIRD_PARTY_NOTICES.md: attribution for PLCrashReporter (MIT), AOSP
tombstone.proto (Apache-2.0), and google/protobuf runtime (BSD-3).

+++ android/src/main/kotlin/co/rivium/trace/sdk/flutter/RiviumTraceFlutterSdkPlugin.kt
@@ -67,7 +67,7 @@ class RiviumTraceFlutterSdkPlugin : FlutterPlugin, MethodChannel.MethodCallHandl
             anrTimeoutMs = (call.argument<Int>("anrTimeoutMs") ?: 5000).toLong(),
             maxBreadcrumbs = call.argument<Int>("maxBreadcrumbs") ?: 20,
             enableOfflineStorage = call.argument<Boolean>("enableOfflineStorage") ?: true,
-            sampleRate = call.argument<Double>("sampleRate") ?: 1.0,
+            sampleRate = (call.argument<Double>("sampleRate") ?: 1.0).toFloat(),
             apiUrl = call.argument<String>("apiUrl") ?: "https://trace.rivium.co"
         )
